Organize frontend modules
