+++ .gitignore
@@ -24,6 +24,6 @@ dist-ssr
 *.sw?
 
 # Will be generated during build process.
-src/assets/*.wasm
-src/assets/wasm_exec.js
+src/wasm/generated/*.wasm
+src/wasm/generated/wasm_exec.js
 wasm/internal/stdlib/stdlib.bundle
