+++ src/app/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
-import "./index.css";
+import "../index.css";
 import "@fontsource/inter";
 import "@fontsource/jetbrains-mono";
 import "@fontsource/jetbrains-mono/800.css";
