+++ src/viewers/shared/lineMapping.ts
