+++ src/viewers/ssa/SSAViewer.tsx
@@ -1,6 +1,6 @@
-import { CodeViewer } from "../CodeViewer.tsx";
+import { CodeViewer } from "../shared/CodeViewer.tsx";
 import React, { useMemo } from "react";
-import { ViewerTitle } from "../ViewerTitle.tsx";
+import { ViewerTitle } from "../shared/ViewerTitle.tsx";
 import { VIEWER_TITLE_HEIGHT } from "../../constants.ts";
 import { SSAFunction } from "../../wasm/protocol.ts";
 import { formatSSA } from "./format.ts";
@@ -8,7 +8,7 @@ import {
   displayLineForSourceLine,
   FormattedAnalysis,
   sourceLineForDisplayLine,
-} from "../lineMapping.ts";
+} from "../shared/lineMapping.ts";
 
 interface SSAViewerProps {
   ssa: SSAFunction[] | string | null;
