+++ src/viewers/ssa/format.ts
@@ -1,5 +1,5 @@
 import { SourcePosition, SSAFunction } from "../../wasm/protocol.ts";
-import { FormattedAnalysis } from "../lineMapping.ts";
+import { FormattedAnalysis } from "../shared/lineMapping.ts";
 
 const formatPosition = (position: SourcePosition) =>
   `${position.line}:${position.column}`;
